pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "specdiff"
version = "0.1.0"
description = "Spec-driven code generation — the spec is the source of truth, code is a build artifact."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Aditya Ramkumar" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click",
    "python-frontmatter",
    "google-genai",
    "google-adk",
    "python-dotenv",
    "pyyaml",
]

[project.urls]
Homepage = "https://github.com/aramkumar/specdiff"
Repository = "https://github.com/aramkumar/specdiff"
Issues = "https://github.com/aramkumar/specdiff/issues"

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
specdiff = "specdiff.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
